+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+<<<<<<< HEAD
+## [2.0.0] - 2025-10-31
+
+### Added
+
+-   Added `typing.Annotated` to all public function signatures.
+-   Fixed all ruff linting errors.
+=======
 # Changelog
 
 All notable changes to this project will be documented in this file.
@@ -27,3 +35,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Various bug fixes and improvements
+>>>>>>> origin/main